Migrate to System.CommandLine for command-line parsing
